pyproject.toml
[project]
name = "purrxelity"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "bcrypt>=5.0.0",
    "docker>=7.1.0",
    "fastapi-cli-slim>=0.0.14",
    "fastapi[all]>=0.120.1",
    "grandalf>=0.8",
    "langchain>=1.0.2",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.1",
    "langchain-google-community>=3.0.0",
    "langchain-google-genai>=3.0.0",
    "langchain-tavily>=0.2.12",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.1",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.3",
    "pypdf>=6.1.3",
    "requests>=2.32.5",
    "sqlalchemy[asyncio]>=2.0.44",
    "sse-starlette>=3.0.2",
]
